p4i integration Integrate 102955 from 3 to main: Integrate 102954 from 3.2 to 3: warning free code..
[git-p4: depot-paths = "//depot/qt/main/": change = 102956]

+++ src/kernel/qtextengine.cpp
@@ -741,7 +741,7 @@ static void calcLineBreaks(const QString &str, QCharAttributes *charAttributes)
 	    // ### handle complex case
 	    ncls = QUnicodeTables::LineBreak_ID;
 	}
-	int brk = charAttributes[i].charStop ? breakTable[cls][ncls] : Pbk;
+	int brk = charAttributes[i].charStop ? breakTable[cls][ncls] : (int)Pbk;
 	if (brk == Ibk)
 	    charAttributes[i].softBreak = (lineBreakClass(uc[i-1]) == QUnicodeTables::LineBreak_SP);
 	else
